refactor: context api 방식으로 header 컴포넌트 수정
