+++ src/context/components/header/HeaderContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const HeaderContext = createContext();
