+++ src/context/components/header/useHeader.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { HeaderContext } from './HeaderContext';
+
+export const useHeader = () => {
+  return useContext(HeaderContext);
+};
